+++ R/MaestroPipelineList.R
@@ -45,8 +45,29 @@ MaestroPipelineList <- R6::R6Class(
           self$MaestroPipelines <- append(self$MaestroPipelines, .x)
         })
       }
+      invisible()
     },
 
+    #' @description
+    #' Update pipelines in a list
+    #' @param MaestroPipelines list of MaestroPipelines
+    #' @return invisible
+    update_pipelines = function(MaestroPipelines = NULL) {
+      pipe_names <- self$get_pipe_names()
+      if ("MaestroPipeline" %in% class(MaestroPipelines)) {
+        pipe_to_update_name <- MaestroPipelines$get_pipe_name()
+        idx_to_update <- which(pipe_names == pipe_to_update_name)
+        self$MaestroPipelines[[idx_to_update]] <- MaestroPipelines
+      } else {
+        purrr::walk(MaestroPipelines, ~{
+          pipe_to_update_name <- .x$get_pipe_name()
+          idx_to_update <- which(pipe_names == pipe_to_update_name)
+          self$MaestroPipelines[[idx_to_update]] <- .x
+        })
+      }
+      invisible()
+    },
+    
     #' @description
     #' Get names of the pipelines in the list arranged by priority
     #' @return character
@@ -326,17 +347,16 @@ MaestroPipelineList <- R6::R6Class(
         purrr::map(...)
       }
 
-      if (!is.null(cores)) {
-        if (cores < 1 || (cores %% 1) != 0) cli::cli_abort("`cores` must be a positive integer")
-        if (cores > 1) {
-          tryCatch({
-            rlang::check_installed("furrr")
-            mapper_fun <- function(...) {
-              furrr::future_map(..., .options = furrr::furrr_options(stdout = TRUE, seed = NULL))
-            }
-          }, error = \(e) {
-            cli::cli_warn("{.pkg furrr} is required for running on multiple cores.")
-          })
+      if (cores > 1) {
+        mapper_fun <- function(...) {
+          furrr::future_map(
+            ..., 
+            .options = furrr::furrr_options(
+              packages = "maestro",
+              stdout = FALSE, 
+              seed = NULL
+            )
+          )
         }
       }
 
@@ -360,45 +380,50 @@ MaestroPipelineList <- R6::R6Class(
                 .input = .input, 
                 run_id = run_id, 
                 input_run_id = input_run_id,
-                lineage = lineage
+                lineage = lineage,
+                depth = depth
               )
             )
           )
         }, error = \(e) {
-          invisible()
+          return(pipe)
         })
         .input <- pipe$get_returns()
         out_names <- network$to[network$from == pipe$get_pipe_name()]
-        if (pipe$get_status_chr() %in% c("Error", "Not Run")) return(invisible())
-        if (length(out_names) == 0) return(invisible())
+        if (pipe$get_status_chr() %in% c("Error", "Not Run")) {
+          return(pipe)
+        }
+        if (length(out_names) == 0) return(pipe)
         lineage <<- append(lineage, pipe$get_pipe_name())
         
         for (i in out_names) {
           pipe <- self$get_pipe_by_name(i)
-          prepend <- paste0(rep("  ", times = depth), "|-")
           run_pipe(
             pipe,
             .input = .input,
             depth = depth,
-            cli_prepend = cli::format_inline(prepend),
             run_id = run_id,
             input_run_id = run_id,
             lineage = lineage
           )
         }
-      }
 
+        pipe
+      }
+      
       # Run the pipelines
       lineage <- NULL
-      mapper_fun(
+      run_res <- mapper_fun(
         pipes_to_run,
         purrr::safely(~{
-          run_pipe(.x)
+          pipe <- run_pipe(.x)
           lineage <<- NULL
+          pipe
         }, quiet = TRUE)
       )
-
-      invisible()
+      
+      purrr::map(run_res, "result") |> 
+        purrr::list_flatten()
     }
   ),